reuse page component
